refactor: move trace tests to artillery package
